requirements.txt
# Core
openai
requests
python-dotenv
pyyaml
numpy

# Visualization
matplotlib
seaborn
pillow

# ML / Embeddings
transformers
torch
sentence-transformers
sentencepiece
protobuf

# Optional / Cloud
boto3
